# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "VK TPA Website API Documentation",
        "title": "TPA-Website",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/flight-transaction": {
            "post": {
                "description": "Add a new flight transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight Transaction"
                ],
                "summary": "Add flight transaction",
                "parameters": [
                    {
                        "description": "Flight Transaction",
                        "name": "flightTransaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight Transaction created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flight-transaction/flight-cart": {
            "post": {
                "description": "Add a new flight cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight Cart"
                ],
                "summary": "Add flight cart",
                "parameters": [
                    {
                        "description": "Flight Cart",
                        "name": "flightCart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight Cart created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flight-transaction/flight-cart/{id}": {
            "get": {
                "description": "Get flight cart by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight Cart"
                ],
                "summary": "Get flight cart by user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight Cart found successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flight-transaction/user/history/{id}": {
            "get": {
                "description": "Get history flight transaction by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight Transaction"
                ],
                "summary": "Get history flight transaction by user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight Transaction found successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flight-transaction/user/ongoing/{id}": {
            "get": {
                "description": "Get on going flight transaction by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight Transaction"
                ],
                "summary": "Get on going flight transaction by user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight Transaction found successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flight-transaction/user/total/{id}": {
            "get": {
                "description": "Get total flight transaction by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight Transaction"
                ],
                "summary": "Get total flight transaction by user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight Transaction found successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flight-transaction/user/{id}": {
            "get": {
                "description": "Get flight transaction by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight Transaction"
                ],
                "summary": "Get flight transaction by user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight Transaction found successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}